requirements.txt
Django
django-simple-history
django-auditlog
django-import-export
pandas
openpyxl
pyxlsb
